Add files via upload
Improved models merged contains all of the improved models merged, resulting in one plot with all the different ROC curves for the different models in the non-galaxy galaxy classification. The with ZM and without ZM improved models for galaxy and non galaxy classifiers are some of the improvements performed on the models including the original code and the last used improved model.