Add files via upload
These files contain the adaptions to the code as given by the paper. The files starting with ViT and ZM are the original files with improvement to code trials. The Improved models merged is the file that contains all the improved final values in order to plot it all in one ROC curve plot.